Initial commit of deviceconfiguration update
Moving the files around to support importing and wrapper .py scripts just
like other applications. Eventually moving towards Python Packaging and
command line functionality.
